core/local/chokidar: Extract analyseDocEvents() from analysis
